test-api/school_management_test/requirements.txt
pytest==7.4.3
requests==2.31.0
pymysql==1.1.0
PyJWT==2.8.0
pytest-html==4.1.1
pytest-xdist==3.5.0
allure-pytest==2.13.2
pyyaml==6.0